build system fixes, mostly for mingw32 cross and win32
git-svn-id: svn://tug.org/texlive/trunk/Build/source@17166 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/kpathsea/dir.c
@@ -43,7 +43,7 @@ kpathsea_dir_p (kpathsea kpse, const_string fn)
   if (KPATHSEA_DEBUG_P (KPSE_DEBUG_STAT)) {
     if (fa == 0xFFFFFFFF) {
       fprintf(stderr, "failed to get file attributes for %s (%d)\n",
-	      fn, GetLastError());
+	      fn, (int)(GetLastError()));
     } else {
       fprintf(stderr, "path %s %s a directory\n",
 	      fn , (fa & FILE_ATTRIBUTE_DIRECTORY) ?
